+++ editor/editor-core/src/plugins/caption/__tests__/visual-regression/alignment.ts
@@ -64,7 +64,8 @@ describe('Snapshot Test: Caption with media', () => {
         await initEditor(page, mediaWithCaptionAlignStart);
       });
 
-      it('show the caption with align end alignment', async () => {
+      // FIXME: This test was automatically skipped due to failure on 8/26/2021: https://product-fabric.atlassian.net/browse/ED-13678
+      it.skip('show the caption with align end alignment', async () => {
         await initEditor(page, mediaWithCaptionAlignEnd);
       });
     });