+++ elements/user-picker/src/__tests__/unit/components/AddOptionAvatarSpec.tsx
@@ -1,5 +1,5 @@
 import EmailIcon from '@atlaskit/icon/glyph/email';
-import { N200 } from '@atlaskit/theme/colors';
+import { N500 } from '@atlaskit/theme/colors';
 import { shallow } from 'enzyme';
 import React from 'react';
 import {
@@ -14,15 +14,15 @@ describe('AddOptionAvatar', () => {
   it('should render email Icon', () => {
     const component = shallowAddOptionAvatar({
       label: 'Invite',
-      size: 'small',
+      isLozenge: false,
     });
 
     const inviteIcon = component.find(EmailIcon);
     expect(inviteIcon).toHaveLength(1);
     expect(inviteIcon.props()).toMatchObject({
       label: 'Invite',
-      size: 'small',
-      primaryColor: N200,
+      size: 'medium',
+      primaryColor: N500,
     });
   });
 });