+++ elements/user-picker/src/components/AddOptionAvatar.tsx
@@ -1,57 +1,34 @@
 import EmailIcon from '@atlaskit/icon/glyph/email';
-import MentionIcon from '@atlaskit/icon/glyph/mention';
-import { N40, N50, N200 } from '@atlaskit/theme/colors';
+import { N40, N500 } from '@atlaskit/theme/colors';
 import React from 'react';
 import styled from 'styled-components';
 
-const AddOptionAvatarWrapper = styled.span`
-  color: black;
-  padding: 2px;
-
-  background-color: ${N50};
-  border-radius: 50%;
-`;
-
 const EmailAvatarWrapper = styled.span`
   color: black;
-  padding: 2px;
+  padding: ${(props: { isLozenge?: boolean }) => (props.isLozenge ? 0 : 4)}px;
 
   background-color: ${N40};
   border-radius: 50%;
+  display: flex;
+  align-items: center;
 `;
 
 export type AddOptionAvatarProps = {
   label: string;
-  size?: 'small' | 'medium' | 'large';
-  suggestion?: boolean;
-  invalidOption?: boolean;
+  isLozenge?: boolean;
 };
 
 export const AddOptionAvatar: React.FunctionComponent<AddOptionAvatarProps> = ({
-  size,
+  isLozenge,
   label,
-  suggestion,
-  invalidOption,
 }) => {
-  if (invalidOption) {
-    return (
-      <AddOptionAvatarWrapper>
-        <MentionIcon
-          label={label}
-          size={size}
-          primaryColor={suggestion ? N200 : 'white'}
-        />
-      </AddOptionAvatarWrapper>
-    );
-  }
-
   return (
-    <EmailAvatarWrapper>
-      <EmailIcon label={label} size={size} primaryColor={N200} />
+    <EmailAvatarWrapper isLozenge={isLozenge}>
+      <EmailIcon
+        label={label}
+        size={isLozenge ? 'small' : 'medium'}
+        primaryColor={N500}
+      />
     </EmailAvatarWrapper>
   );
-};
-
-AddOptionAvatar.defaultProps = {
-  size: 'large',
 };