+++ elements/user-picker/src/components/EmailOption.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
+import { B400, N200, N800 } from '@atlaskit/theme/colors';
 import { Email } from '../types';
 import { AddOptionAvatar } from './AddOptionAvatar';
-import { AvatarItemOption } from './AvatarItemOption';
+import { AvatarItemOption, TextWrapper } from './AvatarItemOption';
 import { messages } from './i18n';
 import { EmailValidationResponse } from './emailValidation';
 
@@ -28,34 +29,42 @@ export class EmailOption extends React.PureComponent<EmailOptionProps> {
         }
       : this.props.email.lozenge;
 
-  private renderOption = (
-    label: string,
-    validity: EmailValidationResponse,
-    suggestion?: boolean,
-  ) => (
+  private renderPrimaryText = () => {
+    const {
+      email: { id },
+    } = this.props;
+
+    return (
+      <TextWrapper key="name" color={this.props.isSelected ? B400 : N800}>
+        {id}
+      </TextWrapper>
+    );
+  };
+
+  private renderSecondaryText = (label: string) => {
+    return (
+      <TextWrapper color={this.props.isSelected ? B400 : N200}>
+        {label}
+      </TextWrapper>
+    );
+  };
+
+  private renderOption = (label: string) => (
     <AvatarItemOption
-      avatar={
-        <AddOptionAvatar
-          suggestion={suggestion}
-          invalidOption={suggestion && validity !== 'VALID'}
-          label={label}
-        />
-      }
-      primaryText={this.props.email.id}
-      secondaryText={label}
+      avatar={<AddOptionAvatar label={label} />}
+      primaryText={this.renderPrimaryText()}
+      secondaryText={this.renderSecondaryText(label)}
       lozenge={this.getLozengeProps()}
     />
   );
 
   render() {
-    const { label, emailValidity, email } = this.props;
+    const { label, emailValidity } = this.props;
     return label !== undefined ? (
-      this.renderOption(label, emailValidity, email.suggestion)
+      this.renderOption(label)
     ) : (
       <FormattedMessage {...getAddEmailMessage(emailValidity)}>
-        {(label) =>
-          this.renderOption(label as string, emailValidity, email.suggestion)
-        }
+        {(label) => this.renderOption(label as string)}
       </FormattedMessage>
     );
   }